Remove utils.cpp
