+++ src/tldr.cpp
@@ -1,5 +1,4 @@
 #include "tldr.h"
-#include "utils.h"
 
 #include <iostream>
 #include <string>
@@ -102,6 +101,16 @@ std::string getUrlForArg(std::string const& arg)
     return getUrlForArgAndPlatform(arg, "common");
 }
 
+void replaceAll(std::string& str, std::string const& from, std::string const& to)
+{
+    if (from.empty()) return;
+
+    size_t start_pos = 0;
+    while((start_pos = str.find(from, start_pos)) != std::string::npos) {
+        str.replace(start_pos, from.length(), to);
+        start_pos += to.length();
+    }
+}
 
 // =====================================
 // Curl Fetching.
