--- src/utils.h
@@ -1,3 +1,0 @@
-#include <string>
-
-void replaceAll(std::string& str, std::string const& from, std::string const& to);
